ref(prettier): move rules into own directory
