+++ src/prettier/index.ts
@@ -1,13 +1,7 @@
-import {
-  apply,
-  chain,
-  mergeWith,
-  Rule,
-  Tree,
-  url
-} from '@angular-devkit/schematics';
-import { installDependencies, PackageJson } from '../lib';
+import { chain, Rule } from '@angular-devkit/schematics';
+import { installDependencies } from '../lib';
 import { patchTsLintConfiguration } from './rules';
+import { configurePrettier, registerPrettier } from './rules/prettier';
 
 export default function(): Rule {
   return chain([
@@ -18,21 +12,4 @@ export default function(): Rule {
     registerPrettier(),
     patchTsLintConfiguration()
   ]);
-}
-
-export function registerPrettier(): Rule {
-  return (tree: Tree) => {
-    const packageJson = new PackageJson(tree.read('package.json'));
-
-    packageJson.setScript(
-      'format',
-      'prettier --write "**/*.{js,json,css,scss,md,ts,html}"'
-    );
-
-    tree.overwrite('package.json', packageJson.stringify());
-  };
-}
-
-export function configurePrettier(): Rule {
-  return mergeWith(apply(url('./templates'), []));
 }
