+++ src/prettier/rules/prettier/configure-prettier.ts
@@ -1,0 +1,4 @@
+import { apply, mergeWith, Rule, url } from '@angular-devkit/schematics';
+export function configurePrettier(): Rule {
+  return mergeWith(apply(url('./templates'), []));
+}
