+++ src/prettier/rules/prettier/index.ts
@@ -1,0 +1,2 @@
+export * from './configure-prettier';
+export * from './register-prettier';
